feat: add basic container starting functionality
`docker-api` exposes functionality to allow interaction with the host
system's Docker API, meaning we can use it to create, start and stop
containers.

This can be used to start the new tag, re-route the traffic, and then
stop the other tag to ensure no downtime.

This commit adds a basic function that will connect to the Docker API,
create a container (using a simple echo server image) and then start the
container.