feat: allow containers to start on random ports
The issue with the containers at the moment is that you must specify the
ports you'd like them to run on, meaning you must know a free port at
the time of creation. Instead, we can just start the container on port
0 and allow the operating system to choose a free port.

This means we can just "start a container" and then check what port it
is running on to decide where to send traffic.