feat: enable additional container creation
Add a hacky `/spawn` route to the load balancer that will create a new
instance of the container and add it to the load balancer's working
state. This means a new container can be brought up for processing while
requests are flying through, and once it has started up it will begin
having traffic routed to it.