feat: allow load balancer to find new images
In order for the load balancer to find new images to run, it needs to
periodically check the Docker registry for new tags. This can be done by
running a separate Tokio task alongside the server itself, which will
fetch all the tags from the registry and compare them to the current
one.

Note, this only works for a specific tag format currently, and doesn't
make any changes upon finding a newer version. It also assumes that the
repository is public, and is hosted on Docker Hub. Other providers are
not supported currently.