feat: add support for configuration files
Passing the configuration for `f2` by command line is a little annoying,
and means it cannot be version controlled. This adds support for config
in the TOML format, which is nice and readable.

This also allows the Docker registry and authentication parameters
(username and password) to be supplied to the tag checker, meaning it
can work across different providers such as GCR, Quay and the default
Docker Hub.

The default location for configuration is `f2.toml`, although this can
be changed with the `--config` flag.