feat: swap to handwritten Docker API library
`cargoship` is a handwritten Docker API library custom to this project,
and currently only exists locally. In the next commit, it will be
inlined into the project as it is a little rough around the edges and
does not provide much functionality.

This allows everything to be queried as expected, including pulling the
images and starting containers with the correct port mappings.