fix: allow fetching image tags
Image tags are failing to be pulled from the Docker Registry as it is
requesting for the image name itself, instead of including the
repository at the start. This means it is requesting `basic-app` instead
of `alexanderjackson/basic-app`.

This probably works for larger, more stable images like `nginx`, but not
for our toy images.

Instead, we should be using the fully qualified name when requesting our
scope and listing the tags.

Additionally, a failure to fetch tags (such as a 401 Unauthorized or
network issues) should not cause the task running this job to fail, so
we should handle and display the error for debugging purposes but
continue our attempts.