feat: wait for image pulls to complete
When running with an unpinned image tag (by leaving it empty), `f2` will
automatically try and use the latest image in the registry. If it does
not exist, it will attempt to pull the image before starting containers.
This pull process obviously also happens when a new tag is available and
`f2` wants to start up new containers.

However, this bit of code was simply making the request to the Docker
host and then continuing on, without waiting for the pull to actually
complete. This is because it is done through a streaming response, where
the host returns the ongoing progress to the client. This means we were
requesting a pull, but then trying to create a container before we had
the image.

Instead, we should `await` on the body being finished. Once this has
completed, we will have successfully pulled the image.

This change:
* Makes sure we finish pulling a Docker image before returning from the
  function