refactor: simplify port mapping logic
The logic to go from a `HashMap<u16, u16>` to the correct setup for the
container is a little unwieldy, mostly because it does everything
through iterators and duplicates some logic. This also means we iterate
the `HashMap` twice instead of the once we actually need.

Instead, we can just use a `for_each` and have mutable `HashMap`s to
prevent the iterators. This also means we don't have the shared logic.

This change:
* Simplifies the port mapping logic in the Docker API handler