feat: support host-based reverse proxying
Most web servers run something like `nginx` in front of them, which
allows them to have multiple downstream servers and decide which one to
send traffic to based on some rules. For example, `example.com` and
`website.app` might route to the same IP address, with `nginx` on the
server deciding which service to route to.

For `f2` to be useful, this will be a needed addition. Even for
`opentracker`, we will need to route `/api` to the backend and
everything else to the frontend.

This change:
* Refactors the way configuration works to make it YAML based
* Allows for multiple services to be running, discriminating where to
  send the request based on the `host` option
* Runs the automated Docker registry poller for each service