feat: add utility function for starting on a port
The `main` function currently has to bind a `TcpListener` and do some
socket handling, which feels a bit odd. It might be easier to provide a
function to start the load balancer on a given port and have that do the
necessary setup before calling the normal `start` function.

This change:
* Adds `start_on_port`, which takes a port number and binds the
  `TcpListener` before calling the regular `start` function