feat: combine account and app to image
Docker works better on the concept of an "image", which can be made up
of just an application (such as `nginx`) or an account alongside the
application name (such as `alexanderjackson/flask-app`).

This should allow side containers like `postgres` to be started without
trying to pull from the defined repository account.

This change:
* Removes the `repository_account` property and encourages the use of
  `image` instead as this can pull Docker-native images