feat: allow environment variables for services (#4)
Currently, you can define a Postgres container on a given tag and it
will start up before immediately crashing. This is because the runtime
expects a `POSTGRES_PASSWORD` to be defined but the API is not providing
one to the container.

Instead, we can allow an optional environment to be passed alongside the
service definition and pass this down to the `CreateContainer` object
that is sent over to Docker. This allows the container to start
correctly.

This change:
* Adds an optional `environment` section to the service configuration
* Passes this down when we create the container

Closes #1.