refactor: remove auto-update functionality
The auto-update functionality requires constantly making requests to
Docker's API (the public one) and a specific tagging format. Instead, we
should probably support remote configuration with some form of webhook
so that clients can write a new configuration and then notify the
controller to reconcile.

This also has the nice benefit of getting rid of a Git dependency as
well as halving the size of the tree.

This change:
* Removes the auto-updating functionality