feat: swap to `color-eyre`
So, may have spent 2 days trying to debug `glibc` errors when in fact
all that was happening was a very generic "could not find file" error
when trying to run `f2` in a container as it had no configuration file.

Let's swap to something more fancy for better (and clearer) error
messages.

This change:
* Removes `anyhow`
* Adds `color-eyre`
* Updates the usages