feat: get Docker working locally (#12)
Due to running on an M2 locally (and thus Apple Silicon), the standard
Dockerfile doesn't seem to work and generally fails with linker errors
around the `ring` library. Adding some extra environment variables
works, so hopefully we can add these for CI as well and produce an image
that works for both.

Additionally, we're reinstalling everything on every file change which
isn't very good caching.

This change:
* Gets the Docker build working locally
* Moves the `apk` and `rustup` invocations earlier for better caching