feat: allow proxying in Docker
`f2` doesn't currently work it is running inside Docker itself, since it
tries to proxy everything to `localhost` and the port binding that's
exposed to the local machine.

However, we'll be on the same Docker network as the running containers
we want to proxy to and they'll be exposing a port (that we then expose
on a random port to the local machine). Thus, all we need to do is grab
those IP addresses and make requests to them instead.

We already get some of the network details, so this is an easy addition
to that request.

This change:
* Grabs the IP address of the running containers
* Proxies requests to those instead
* Ignores the exposed ports for now, these can probably be tidied up in
  future or `f2` can run in Docker/non-Docker modes
* Updates the tests to use `LOCALHOST` as we now deal in `SocketAddrV4`
  a lot more