refactor: stop searching for bound ports
Now that everything is in the same Docker network, we don't actually
need to expose any ports to the host machine. Docker will always give us
a unique IP address for the container (until we run out) and thus we can
just use this along with the host port defined in the YAML.

This removes a lot of the odd complexity around exposing ports initially
and then trying to retrieve them.

This change:
* Removes the functionality, assuming we will always be running `f2`
  inside Docker itself