feat: add config reconciliation
In the future, the ideal workflow will be that:

* Some code is pushed up to a PR
* The change is merged and a new Docker image is built
* Another bit of code automatically updates the config in the S3 bucket
  and tells `f2` to reconcile the changes
* `f2` fetches the latest configuration and works out what has changed
* It notices an image has been updated, so it starts the new one and
  terminates the old ones

This will bring us to a state where it can just be running on the EC2
instance and nothing needs to touch it. Deployments will be as fast as
possible and all that needs to happen is code to be written. Services
can be added to the config and `f2` will notice them and deploy them.
Tags will change and services will roll.

This is a rather messy starting point for that. There's a lot to be
worked out here, data is in the wrong place and some of the model is
clearly a bit wonky. But it does work! If you run it locally and change
a tag before telling it to reconcile, it will start the new image and
add it to the load balancer.

Some improvements need to be made:

* Likely to want a proper service registry for names, definitions,
  downstreams, states and such
* There's a lot of shared state currently, the reconciler should not be
  the place to get the reconciliation path from
* The rolling of new tags has little error handling

This change:
* Adds the ability to produce a diff of configuration
* Adds a `Reconciler` to deal with producing diffs and sending them to
  the load balancer
* Mangles lots of code