feat: add service registry
While writing the configuration reconciliation functionality, it became
very difficult to tell what was running for a given service. We don't
even track the running containers yet, so there's no way to terminate
the old versions after the new version has been brought up.

It also had to make some weird decisions about which identifiers to use
where, as some code used the service name and others used the image.

Instead, this adds a `ServiceRegistry` type to handle this. Services can
be defined and containers registered, with the type handling questions
such as "find me the downstreams", "get me the definition of a service"
and "get me the running containers for a service". This also makes it
much easier to test.

This change:
* Adds the type
* Refactors everything to use it
* Updates the tests and adds a few more