feat: remove containers during rolls (#21)
Previously, the load balancer would just start the new container and
then begin sending it requests. It would not however remove the old one
(and would continue sending it requests), meaning that you could have an
older version of it still running.

Additionally, the load balancer didn't feel like quite the right place
to be handling the logic. Ideally it should just know about downstream
clients to send requests to and nothing else, it does not need to
interact with Docker itself.

Instead, the reconciler should house this logic. It can have access to
the service registry and then fetch the latest config, compute the diff
and then apply the changes to the Docker client for the load balancer to
pick up.

This change:
* Moves the roll logic to the reconciler
* Allows it to delete the containers when it has started the new ones
* Updates the tests to account for the new structure