fix: roll correct replica count
As outlined in #22, the reconciler currently only rolls a single new
container when it notices a tag update. This isn't great, as users can
define a custom number of replicas to use. If someone defines 3
containers and then rolls to a new tag, they'll only have 1 running.
Worse, if they roll back, they'll also still just have the 1 container.

This change:
* Starts as many replicas as requested when reconciling tag diffs