feat: handle addition and removal of services
Over time, services may be added or removed from the configuration. In
order to handle these properly, the `Diff` struct needs some additional
variants to determine what's happening. As well as this, the reconciler
needs to understand how to apply those to the state.

Additionally, the state of the definitions was slightly wrong after some
of the reconciler updates, so this makes the necessary changes to
correct that.

This change:
* Adds the new variants
* Handles them in the reconciler