feat: support TLS
Ideally, `f2` would be the only thing running on the instance. Currently
`nginx` is used to terminate the TLS connection, however this is
something that `f2` should be able to do as well.

Unfortunately this requires copy-pasting the service definition for now,
as it needs to be made generic over multiple acceptor types.

This change:
* Adds support for TLS termination in the load balancer