feat: support remote private keys
Currently the private key needs to be passed as an environment variable
on startup, which is somewhat annoying. Every time a new instance is set
up it much be downloaded from 1Password and re-encoded.

Ideally, the instance could just pull the key from a remote location on
startup and then decrypt the secrets.

This change:
* Adds support for remote private keys
* Makes private keys optional, as they're not always required