feat: support all service alterations (#34)
Sometimes, a container may have its tag bumped as well as some new
configuration being added. This causes issues as `f2` will use the old
environment variables for the new image and thus it may fail to start
(which is not handled very gracefully at the moment).

Instead of listening only for tag updates, we can just check for any
changes to the service entirely. This means changes to the port, tag,
replicas and environment will all cause a roll.

`clippy` also didn't like everything starting with the same prefix, so
this renames the variants too.

This change:
* Adds the new `Alteration` variant and removes `TagUpdate`
* Handles the new variant in the reconciler
* Updates the tests to handle it