fix: decrypt only on container creation (#36)
As described in #35, the reconciliation doesn't handle secrets very well
since it will compare the locally decrypted ones to the encrypted ones
in the file and decide that they have changed. It will then start
services with the encrypted values which can cause failures.

Instead, we should keep them decrypted until we need them. This allows
the containers to always get decrypted secrets, but we always compare
against the encrypted versions (which is better for key changes as
well).

This change:
* Adds the ability to fetch the private key to the configuration so it
  can be used in a few places
* Updates the container creation to take a private key and do the
  decryption
* Adds some tests to make sure the decryption behaves properly