refactor: remove unnecessary `collect`
Since the service registry stores a `HashSet` of downstream addresses,
we can't easily pick a random one to go to when proxying requests. This
means we need to `collect` into a `Vec` before choosing, which is a
little annoying.

`indexmap` provides an `IndexSet` type which has the best of both
worlds, we simulataneously cannot have duplicates but also can easily
select a random index.

This change:
* Adds a dependency on `indexmap`
* Updates the service registry to use `IndexSet` over `HashSet`
* Fixes the selection methodology