refactor: abstract Docker away (#43)
The Docker functionality gets in the way of testing fairly frequently,
since this is a container orchestrator. This means the reconciler is
difficult to write tests for as it interacts heavily with creating new
containers or removing existing ones.

Instead of this, we can work towards using a `DockerClient` trait. This
allows us to use a fake implementation in test code that just stores a
bit of state and acts as if it has talked to the Docker API. Production
code can still talk to Docker normally.

This also has the nice benefit of reducing the number of places we
instantiate a Docker client (to just the one).

This change:
* Introduces the trait, regular and fake implementations
* Adds a test to the reconcilier to do some sanity checks