fix: correct alteration handling
The alteration handling part of the reconciler currently only creates
the number of replicas specified in the original definition, not the new
one. This means swapping from 1 replica to 2 just replaces the existing
container instead of ending with 2 containers.

This was only identified due to some unit tests, so it's good that we
can now begin writing them.

This change:
* Corrects the bug
* Adds a unit test to ensure it works as expected now
* Removes some unused code to appease `clippy`