feat: enforce non-zero replica counts
It probably doesn't make much sense to have 0 replicas of a service
(just remove it entirely if you need that) and if we have a constraint
like that then we should aim to enforce it at the API level. Rust has a
`NonZeroU8` type, so let's wrap that and provide a little extra
functionality around it.

This change:
* Adds a `ReplicaCount` type
* Uses it everywhere it needs to be