fix: defaults for environment and volumes
`f2` is failing to start up on the new version since the `volumes`
property doesn't exist. It isn't actually required (plenty of containers
won't have associated volumes) so let's just default to an empty
`HashMap` if we don't find the property.

Turns out we can do this for the `environment` as well, so let's tidy up
the code a little bit.

This change:
* Annotates `environment` and `volumes` with `#[serde(default)]`
* Updates the code to handle all this