feat: begin adding health checks (#54)
Currently, the orchestrator just starts containers. There's no knowledge
of whether it started correctly (and thus we should alert someone, or
begin routing traffic to it) nor whether it has died off for some reason
and now requests are going to fail (at which point we should definitely
alert someone).

The aim here is to add some basic health checking in the future. The
orchestrator should start containers and then periodically request to
them before adding them to the load balancer, to ensure they've started
correctly. From there, we can start health checking them all the time to
make sure they don't silently fall over.

This is yet to be used, so shouldn't have any impact on the server.

This change:
* Adds a `health` module with some basic health checking functionality