feat: upgrade `hyper`
`hyper` has been on version 1.0 for a while now, but the ecosystem has
taken a bit of time to catch up. It's probably in a good state now and
I've done some separate work to support mTLS (which is still disabled
here) which resolved a lot of the issues.

This change:
* Updates `f2` to the latest version of `hyper`
* Adds (disabled) support for mTLS