feat: add mTLS support
Now that we've deployed the upgrade to the latest version of `hyper`, we
can start looking at supporting mTLS for some domains.

The general approach here is just to take an anchor certificate that all
the downstream ones must be signed by (this will be the self-signed root
one) and some domains where we should check for authentication.

This change:
* Adds an optional section to the configuration
* Updates the usages to read the certificate and apply the correct
  protocols per domain