refactor: improve proxy error handling
When moving to `hyper v1`, the error handling in the proxy function
started using `unwrap` everywhere since the error variant of the return
type was `hyper::Error` and this cannot be created.

The underlying `serve_connection` function constrains it to anything
that can be converted to a `Box<Error>` essentially, which
`color_eyre::Report` can be. This required `mutual-tls` to support it,
which alexander-jackson/mutual-tls@d9691d8 did.

This change:
* Bumps the version of `mutual-tls`
* Updates the proxy function to use `color_eyre::Report` and `?`
  everywhere