feat: begin supporting graceful shutdown
When servers roll, we should be more graceful about the way in which we
terminate them otherwise clients will receive errors. In testing of a
simulated slow endpoint, this produced a lot of errors to `oha`.

Once `forkup` had been updated to use `axum`'s graceful shutdown
features this disappeared.

This change:
* Adds the ability for containers to opt-in to graceful shutdowns, where
  the load balancer will use a `SIGTERM` instead of a `SIGKILL`
  initially