fix: stop using volumes (#80)
The Docker documentation specifies:

Volumes are not a good choice if you need to access the files from the
host, as the volume is completely managed by Docker. Use bind mounts if
you need to access files or directories from both containers and the
host.

This might be the reason that we're seeing odd behaviour on the server,
and everything seems to be fine locally so it's worth trying out.

This change:
* Stops using volumes when creating containers