feat: embrace event-driven architecture more
Various bits of functionality are currently reactive, such as the
reconciliation process where old pods are torn down and new ones are
brought up. This is done on requests to the `/reconciliation` endpoint
and allow the `infrastructure` repository to force the deployments to be
updated.

However, a lot of functionality still requires restarts. Whenever the
certificates are rolled for example, we need to restart the instance in
order to load the new ones.

This change:
* Adds an `ipc` module for handling a basic message bus
* Makes the certificate resolver more dynamic, allowing it to reload
  certificates from the underlying store
* Adds various tests for the new functionality
* Removes the `ConfigurationLocation` type in favour of `ExternalBytes`
* Removes the implementation of `Client::new` in favour of
  `Client::default`
* Simplifies the implementation of `read_body` to use `http-body-util`
* Updates `handle_request` to be generic over the body type
* Adds tests for the path matching algorithm
* Improves the logging in the reconciler