feat: support multiple hosts (#88)
In order to expose HyperDX internally and externally, we'll need to
proxy multiple hosts into the same container but with different ports.
Internal clients will need to talk to the 4318 port to export traces,
whereas external clients will need to talk to the 8080 port to handle
regular requests from the frontend.

This change:
* Supports multiple hosts with different port mappings