feat: map requests to HTTP 1.1
Currently, `f2` is capable of reading an HTTP2 request but when
proxying it to the backend, that may not be capable of handling it which
causes it to fail the request.

Instead, let's just map all incoming HTTP2 requests into HTTP 1.1 style
for the time being. In the future, we'll allow these to be proxied
transparently for services that do understand HTTP2.

This change:
* Maps all requests into an HTTP/1.1 style when proxying
* Adds tests for some of this